pyproject.toml
[project]
name = "ai-tax-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Chris Andrasick",email = "chris@andrasick.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "pypdf (>=5.4.0,<6.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "langchain-core (>=0.3.51,<0.4.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "langchain-google-genai (>=2.1.2,<3.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "chromadb (>=0.5.4)",
    "onnxruntime (==1.18.0)",
    "pdfplumber (>=0.11.6,<0.12.0)",
    "html5lib (>=1.1,<2.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "google-search-results>=2.4.2,<3.0.0",
    "numexpr (>=2.10.2,<3.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "textstat (>=0.7.5,<0.8.0)",
    "scipy (>=1.15.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-dotenv = "^0.5.2"

